refactor(models): Extract utility functions to utils.py
