+++ scripts/run.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+uv run -m src.models.main
